+++ NEWS.md
@@ -8,8 +8,10 @@ TODO:
 
 -->
 
-# clusterProfiler 4.5.1.900
+# clusterProfiler 4.5.1.901
 
++ `gson_KEGG()` download latest KEGG and output a GSON object (2022-06-08, Wed)
++ support passing a GSON object to `gseKEGG(organism)` 
 + support passing a GSON object to `enrichKEGG(organism)` (2022-06-06, Mon)
 
 # clusterProfiler 4.5.1